refactor: descomponentizando widgets
